+++ packages/core-js/web/request-idle-callback.js
@@ -1,0 +1,5 @@
+'use strict';
+require('../modules/web.request-idle-callback');
+var path = require('../internals/path');
+
+module.exports = path.requestIdleCallback;